backend/requirements.txt

# Data Handling
pandas
numpy

# Machine Learning & NLP
scikit-learn
nltk
gensim
spacy
joblib
torch
tensorflow
transformers
sentence-transformers
textblob


# Backend
fastapi
uvicorn
pydantic
python-multipart

# File parsing
docx2txt
python-docx
PyPDF2

# Utils
tqdm
python-dotenv
